+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.2 (unreleased)
 
 - Improved support for Mac ARM
+- Fixed error with `dup` and `clone`
 
 ## 0.1.1 (2022-06-12)